+++ build.sh
@@ -1,9 +1,7 @@
-if [ ! -x configure ]; then
-    ./autogen.sh
-fi
-
+#!/bin/sh
+./autogen.sh
 ./configure --prefix=/usr --localstatedir=/var --sysconfdir=/etc --enable-builtin-icons \
 	    --without-config --without-ssl --without-symlinks --enable-htaccess --enable-htpasswd
 
-make -j5 clean
+make clean
 make -j5