+++ src/tdate_parse.c
@@ -32,9 +32,6 @@
 #include <config.h>
 
 #include <ctype.h>
-#ifdef HAVE_MEMORY_H
-#include <memory.h>
-#endif
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -58,9 +55,7 @@ static void pound_case(char *str)
 	}
 }
 
-static int strlong_compare(v1, v2)
-char *v1;
-char *v2;
+static int strlong_compare(const void *v1, const void *v2)
 {
 	return strcmp(((struct strlong *)v1)->s, ((struct strlong *)v2)->s);
 }